Consistent library names on Darwin
Before, we were naming the Darwin shared library libluajit-51, instead
of the libluajit-5.1 that we use on Linux and friends.  This caused
issues with the pkg-config file, since it only lists libluajit-5.1 —
making you always link with the static library on Darwin — and with any
hand-crafted build scripts, since you would have to include
platform-specific logic to decide which library name to link with.

This patch updates the LuaJIT build scripts to use libluajit-5.1 as the
library name on all POSIX platforms, for both the shared and static
library, eliminating these issues.